refactor: separate business logic into services
